new module, moved tests, and deliverapi
