+++ cg/meta/deliver/deliver_ticket.py
@@ -10,9 +10,9 @@ from cg.exc import CgError
 from cg.meta.meta import MetaAPI
 from cg.models.cg_config import CGConfig
 from cg.services.fastq_file_service.fastq_file_service import FastqFileService
-from cg.services.fastq_file_service.utils import (
-    get_concatenated_forward_read_output_path,
-    get_concatenated_reverse_read_output_path,
+from cg.meta.deliver.fastq_path_generator import (
+    generate_forward_concatenated_fastq_delivery_path,
+    generate_reverse_concatenated_fastq_delivery_path,
 )
 from cg.store.models import Case
 
@@ -92,10 +92,10 @@ class DeliverTicketAPI(MetaAPI):
                 continue
             if not dir_path.is_dir():
                 continue
-            forward_output_path = get_concatenated_forward_read_output_path(
+            forward_output_path = generate_forward_concatenated_fastq_delivery_path(
                 fastq_directory=dir_path, sample_name=dir_path.name
             )
-            reverse_output_path = get_concatenated_reverse_read_output_path(
+            reverse_output_path = generate_reverse_concatenated_fastq_delivery_path(
                 fastq_directory=dir_path, sample_name=dir_path.name
             )
             self.fastq_file_service.concatenate(
