+++ cg/services/fastq_file_service/utils.py
@@ -37,30 +37,6 @@ def get_forward_read_fastqs(fastq_directory: Path) -> list[Path]:
     return get_fastqs_by_direction(fastq_directory=fastq_directory, direction=ReadDirection.FORWARD)
 
 
-def get_concatenated_read_output_path(
-    fastq_directory: Path, sample_name: str, direction: int
-) -> Path:
-    return Path(
-        fastq_directory, f"{sample_name}_{direction}{FileExtensions.FASTQ}{FileExtensions.GZIP}"
-    )
-
-
-def get_concatenated_forward_read_output_path(fastq_directory: Path, sample_name: str) -> Path:
-    return get_concatenated_read_output_path(
-        fastq_directory=fastq_directory,
-        sample_name=sample_name,
-        direction=ReadDirection.FORWARD,
-    )
-
-
-def get_concatenated_reverse_read_output_path(fastq_directory: Path, sample_name: str) -> Path:
-    return get_concatenated_read_output_path(
-        fastq_directory=fastq_directory,
-        sample_name=sample_name,
-        direction=ReadDirection.REVERSE,
-    )
-
-
 def get_reverse_read_fastqs(fastq_directory: Path) -> list[Path]:
     return get_fastqs_by_direction(fastq_directory=fastq_directory, direction=ReadDirection.REVERSE)
